+++ ldaplib/ldap_docker_test.go
@@ -3,7 +3,10 @@ package ldaplib
 import (
 	"fmt"
 	"os"
+	"path"
 	"time"
+
+	"github.com/tommi2day/gomodules/test"
 
 	"github.com/go-ldap/ldap/v3"
 	"github.com/ory/dockertest/v3"
@@ -11,55 +14,70 @@ import (
 	"github.com/tommi2day/gomodules/common"
 )
 
-const repo = "docker.io/osixia/openldap"
-const repoTag = "1.5.0"
-const containerTimeout = 120
+const Ldaprepo = "docker.io/bitnami/openldap"
+const LdaprepoTag = "2.6.7"
+const LdapcontainerTimeout = 120
 
-var containerName string
+var ldapcontainerName string
 var ldapContainer *dockertest.Resource
 
-// prepareContainer create an Oracle Docker Container
-func prepareContainer() (container *dockertest.Resource, err error) {
+// prepareContainer create an OpenLdap Docker Container
+func prepareLdapContainer() (container *dockertest.Resource, err error) {
 	if os.Getenv("SKIP_LDAP") != "" {
-		err = fmt.Errorf("skipping ORACLE Container in CI environment")
+		err = fmt.Errorf("skipping LDAP Container in CI environment")
 		return
 	}
+	ldapcontainerName = os.Getenv("LDAP_CONTAINER_NAME")
+	if ldapcontainerName == "" {
+		ldapcontainerName = "ldaplib-openldap"
+	}
+
 	var pool *dockertest.Pool
-	containerName = os.Getenv("CONTAINER_NAME")
 	pool, err = common.GetDockerPool()
 	if err != nil {
-		err = fmt.Errorf("could not connect to Docker: %s", err)
 		return
 	}
-
 	vendorImagePrefix := os.Getenv("VENDOR_IMAGE_PREFIX")
-	repoString := vendorImagePrefix + repo
+	repoString := vendorImagePrefix + Ldaprepo
 
-	fmt.Printf("Try to start docker container for %s:%s\n", repoString, repoTag)
+	fmt.Printf("Try to start docker container for %s:%s\n", repoString, LdaprepoTag)
+	fmt.Println(path.Join(test.TestDir, "docker", "ldap", "certs") + ":/opt/bitnami/openldap/certs:ro")
 	container, err = pool.RunWithOptions(&dockertest.RunOptions{
 		Repository: repoString,
-		Tag:        repoTag,
+		Tag:        LdaprepoTag,
 		Env: []string{
-			"LDAP_ORGANISATION=" + ldapOrganisation,
-			"LDAP_DOMAIN=" + LdapDomain,
-			"LDAP_BASE_DN=" + LdapBaseDn,
+
+			"LDAP_PORT_NUMBER=1389",
+			"LDAP_LDAPS_PORT_NUMBER=1636",
+			"BITNAMI_DEBUG=true",
+			"LDAP_ROOT=" + LdapBaseDn,
+			"LDAP_ADMIN_USERNAME=admin",
 			"LDAP_ADMIN_PASSWORD=" + LdapAdminPassword,
-			"LDAP_CONFIG_PASSWORD=" + LdapConfigPassword,
-			"LDAP_TLS_VERIFY_CLIENT=never",
+			"LDAP_CONFIG_ADMIN_ENABLED=yes",
+			"LDAP_CONFIG_ADMIN_USERNAME=config",
+			"LDAP_CONFIG_ADMIN_PASSWORD=" + LdapConfigPassword,
+			// "LDAP_SKIP_DEFAULT_TREE=yes",
+			// "LDAP_CUSTOM_LDIF_DIR=/bootstrap/ldif",
+			"LDAP_CUSTOM_SCHEMA_DIR=/bootstrap/schema",
+			"LDAP_ADD_SCHEMAS=yes",
+			"LDAP_EXTRA_SCHEMAS=cosine,inetorgperson,nis",
+			"LDAP_ALLOW_ANON_BINDING=yes",
+			"LDAP_ENABLE_TLS=yes",
+			"LDAP_TLS_CERT_FILE=/opt/bitnami/openldap/certs/ldap.example.local-full.crt",
+			"LDAP_TLS_KEY_FILE=/opt/bitnami/openldap/certs/ldap.example.local.key",
+			"LDAP_TLS_CA_FILE=/opt/bitnami/openldap/certs/ca.crt",
+			"LDAP_TLS_VERIFY_CLIENTS=never",
 		},
-		Hostname: containerName,
-		Name:     containerName,
-		// ExposedPorts: []string{"389", "636"},
-		/*
-			PortBindings: map[docker.Port][]docker.PortBinding{
-				"389": {
-					{HostIP: "0.0.0.0", HostPort: fmt.Sprintf("%d", ldapPort)},
-				},
-				"636": {
-					{HostIP: "0.0.0.0", HostPort: fmt.Sprintf("%d", sslport)},
-				},
-			},
-		*/
+
+		Mounts: []string{
+			test.TestDir + "/docker/ldap/ldif:/bootstrap/ldif:ro",
+			test.TestDir + "/docker/ldap/schema:/bootstrap/schema:ro",
+			test.TestDir + "/docker/ldap/entrypoint:/docker-entrypoint-initdb.d",
+			test.TestDir + "/docker/ldap/certs:/opt/bitnami/openldap/certs:ro",
+		},
+
+		Hostname: ldapcontainerName,
+		Name:     ldapcontainerName,
 	}, func(config *docker.HostConfig) {
 		// set AutoRemove to true so that stopped container goes away by itself
 		config.AutoRemove = true
@@ -71,10 +89,10 @@ func prepareContainer() (container *dockertest.Resource, err error) {
 		return
 	}
 
-	pool.MaxWait = containerTimeout * time.Second
-	host, port := common.GetContainerHostAndPort(container, "389/tcp")
-	dialURL := fmt.Sprintf("ldap://%s:%d", host, port)
-	fmt.Printf("Wait to successfully connect to Ldap with %s (max %ds)...\n", dialURL, containerTimeout)
+	pool.MaxWait = LdapcontainerTimeout * time.Second
+	myhost, myport := common.GetContainerHostAndPort(container, "1389/tcp")
+	dialURL := fmt.Sprintf("ldap://%s:%d", myhost, myport)
+	fmt.Printf("Wait to successfully connect to Ldap with %s (max %ds)...\n", dialURL, LdapcontainerTimeout)
 	start := time.Now()
 	var l *ldap.Conn
 	if err = pool.Retry(func() error {
@@ -84,11 +102,20 @@ func prepareContainer() (container *dockertest.Resource, err error) {
 		fmt.Printf("Could not connect to LDAP Container: %s", err)
 		return
 	}
+	/*
+		if err = pool.Retry(func() error {
+			err = l.Bind("cn=admin,"+LdapBaseDn, LdapAdminPassword)
+			return err
+		}); err != nil {
+			fmt.Printf("Could not login to LDAP Container: %s", err)
+			return
+		}
+	*/
 	_ = l.Close()
-	// wait 15s to init container
-	time.Sleep(15 * time.Second)
 	elapsed := time.Since(start)
 	fmt.Printf("LDAP Container is available after %s\n", elapsed.Round(time.Millisecond))
+	// wait 15s to init container
+	time.Sleep(15 * time.Second)
 	err = nil
 	return
 }